reformatted - no other changes

+++ app/src/main/java/com/example/osu_bathroom_app/ui/AddReviewFragment.java
@@ -82,7 +82,7 @@ public class AddReviewFragment extends Fragment
         long n = 0;
         float rating = bar.getRating();
         String review = reviewText.getText().toString();
-        long l = (long) globalClass.getUserId();
+        long l = globalClass.getUserId();
         Log.i("BID", "" + bathroomId);
         FirebaseDatabase.getInstance().getReference().child("NumId").get().addOnCompleteListener(new OnCompleteListener<DataSnapshot>()
         {
